fixed fold change
git-svn-id: https://svn-bccfr.uis.georgetown.edu/svn/gdoc/webapp/trunk@300 7e94895d-eddb-4489-9ea7-1067dfd99f16

+++ grails-app/controllers/KmGeneExpCommand.groovy
@@ -5,7 +5,7 @@ class KmGeneExpCommand {
 	AnalysisType requestType = (AnalysisType.KM_GENE_EXPRESSION)
 	List reporters
 	String currentReporter
-	Integer foldChange
+	Double foldChange
 	Integer geAnalysisId
 	String geneName
 	String dataFile
